Suppport transpilation errors logging
The commit removes all sources with the idea to use only a
main.js file. The file gets transpiled and Git should be
used in order to create different examples (in the form
of branches) using just the main as single file. This will
allow an easier way to keep all code with less files around.